Adicionado tela de detalhes da obra de arte feito pela Andressa

+++ app/src/main/java/com/example/appcultural/entities/Art.kt
@@ -1,7 +1,7 @@
 package com.example.appcultural.entities
 
 data class Art(
-    public val id: Long,
+    public val id: Int,
     public val name: String,
     public val publishDate: Int,
     public val description: String,
